pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cdlreq"
version = "0.1.0"
description = "A Python library and CLI for managing medical software requirements and specifications"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.0",
    "click>=8.0",
    "openpyxl>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.10",
    "black>=22.0",
    "flake8>=5.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
]
test = [
    "pytest>=6.0",
    "pytest-cov>=4.0", 
    "pytest-mock>=3.10",
]

[project.scripts]
cdlreq = "cdlreq.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["cdlreq*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true